refactor(reports): decompose ReportsService into Receivables/Revenue/Operational report services
Behavior-preserving, pure-read. Facade keeps its public surface + the thin
AccountingService pass-throughs + 2 scope helpers; the 9 heavy report methods
moved VERBATIM into 3 internally-constructed sub-services (prisma-only). Internal
construction keeps the facade's 2-arg ctor so all 14 `new ReportsService(...)`
spec sites + module + consumers are untouched. Aging-bucket labels left as-is
(trap). tests green, tsc 0, lint 0.
